requirements.txt
psycopg2-binary
pandas
plotly
